backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
supabase>=2.4.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pandas>=2.2.0
scikit-learn>=1.4.0
reportlab>=4.1.0
PyPDF2>=3.0.1
pytesseract>=0.3.10
Pillow>=10.0.0
python-multipart>=0.0.9
